% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aria.R, R/aria_server.R, R/aria_ui.R
\name{aria}
\alias{aria}
\alias{aria_server}
\alias{aria_ui}
\title{Aria web application}
\usage{
aria(
  max_file_size = 100,
  options_port = 8888,
  options_launch_browser = TRUE,
  version = "test"
)

aria_server(version = "test")

aria_ui(version)
}
\arguments{
\item{max_file_size}{The maximum file size in megabytes.}

\item{options_port}{The TCP port that the application should listen on, usually 8888.}

\item{options_launch_browser}{If true, the system's default web browser will be
launched automatically after the app is started.}

\item{version}{Use "test" to run locally, "online" to run in the server
or "desktop" to build the desktop version.}
}
\value{
A web application.
}
\description{
This is the web application using the sistec package. It was created to ease the
work using the package, but you can have the same results reading the
files (read_rfept()), comparing the results with (compare_sistec()) and write
the outputs (write_output()).
}
\examples{
\dontrun{
library(sistec)

# use this to run aria in your R environment
aria()

# use this if you want to run the aria in a server
ui <- aria_ui()
server <- aria_server(version = "online")
shiny::shinyApp(ui, server)
}

}
