% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aria_server.R
\name{aria_server}
\alias{aria_server}
\title{ARIA's server}
\usage{
aria_server(version = "test")
}
\arguments{
\item{version}{Use "test" to run locally, "online" to in the server
or "desktop" to build the desktop version.}
}
\description{
This function calls the server from ARIA
}
