% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aria_ui.R
\name{aria_ui}
\alias{aria_ui}
\title{ARIA's user interface}
\usage{
aria_ui(test_mode = FALSE)
}
\arguments{
\item{test_mode}{A logical to call a test or a production app.#'}
}
\description{
This function calls the UI from ARIA
}
