% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sistec.R
\name{compare_sistec}
\alias{compare_sistec}
\title{Comparison between Sistec and a student registration dataset}
\usage{
compare_sistec(
  sistec,
  student_registration,
  output_path = NULL,
  output_folder_name = "Sistec"
)
}
\arguments{
\item{sistec}{The folder's path to Sistec files or the Sistec data frame
read by \code{sistec::read_sistec()}function.}

\item{student_registration}{The folder's path to students reagistration datasets or a
data frame read by \verb{sistec::read_*()}functions.}

\item{output_path}{Specify a path if you want to save the comparison. The results will
be saved in a folder named by \code{output_folder_name} parameter. The files created will be
saved in xlsx format.}

\item{output_folder_name}{The folder name where the comparison will be saved. If not
specified, the folder name will be "Sistec". This parameter is useless if \code{output_path} was not specified.}
}
\description{
A generic function to compare and save the student situation. This function also
shows inconsistecies in the datasets. You can pass the folder's file path or
a data frame read by \verb{sistec::read_*()} functions.
}
