% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparar_qacademico.R
\name{compare_sistec_qacademico}
\alias{compare_sistec_qacademico}
\title{Comparison beteween sistec and qacademico}
\usage{
compare_sistec_qacademico(sistec_path, qacademico_path)
}
\arguments{
\item{sistec_path}{A list of sistec's files path. This list, normally, comes from the app.}

\item{qacademico_path}{A list of qacademico's files path. This list, normally, comes from the app.}
}
\description{
This function runs the comparison beteween students situation in sistec and qacademico.
}
