% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sistec_complete.R
\name{compare_sistec_qacademico_complete}
\alias{compare_sistec_qacademico_complete}
\title{Comparison between sistec and qacademico}
\usage{
compare_sistec_qacademico_complete(
  sistec_path,
  qacademico_path,
  type = "complete",
  write_output_path = "",
  institute = "IFPE"
)
}
\arguments{
\item{sistec_path}{Path to the sistec files or the sistec dataset read by \code{read_sistec()}.}

\item{qacademico_path}{Path to the qacademico files or the qacademico.
dataset read by \code{read_qacademmico()}.}

\item{type}{Use "complete" to compare.}

\item{write_output_path}{The path where to save the the output in xlsx. If you don't specify the
path, the files won't be saved.}

\item{institute}{The name of your Institute.}
}
\description{
Compare and save the student situation and the inconsistecies in the datasets.
}
