% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparar_tudo.R
\name{output_screen}
\alias{output_screen}
\title{Qacademico screen}
\usage{
output_screen(input_qacademico, input_sistec)
}
\arguments{
\item{input_qacademico}{A list of qacademico's files path. This list, normally, comes from the app.}

\item{input_sistec}{A list of sistec's files path. This list, normally, comes from the app.}
}
\description{
This function runs the comparison beteween sistec and qacademico, make the download of the
results and plot screen results in the app.
}
