% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnp_critics.R
\name{pnp_critics}
\alias{pnp_critics}
\title{Make the pnp critics from sistec files}
\usage{
pnp_critics(students, ciclo)
}
\arguments{
\item{students}{A sistec students data frame from \code{read_sistec_student()}.}

\item{ciclo}{A sistec ciclo data frame from \code{read_ciclo()}.}
}
\value{
A list with the critics
}
\description{
This function gives the pnp critics that can be retrive using sistec information.
}
