% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ciclo.R
\name{read_ciclo}
\alias{read_ciclo}
\title{Read ciclo files from sistec

This function provides support to read ciclo files that can be extarct from sistec website.}
\usage{
read_ciclo(path = "", start = NULL)
}
\arguments{
\item{path}{The folder's path to ciclo files.}

\item{start}{A character with the date to start the comparison. The default is the minimum
value found in the data. The date has to be in this format: "yyyy.semester".
Ex.: "2019.1" or "2019.2".}
}
\description{
Read ciclo files from sistec

This function provides support to read ciclo files that can be extarct from sistec website.
}
