% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_conecta.R
\name{read_conecta}
\alias{read_conecta}
\title{Read Conecta files}
\usage{
read_conecta(path = "", start = NULL)
}
\arguments{
\item{path}{The Conecta file's path.}

\item{start}{A character with the date to start the comparison. The default is the minimum
value found in the data. The date has to be in this format: "yyyy.semester".
Ex.: "2019.1" or "2019.2".}
}
\value{
A data frame.
}
\description{
This function supports two kinds of schemas: from the api and the website. See Details
if you need help to download the Conecta data.
}
