% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_generic_rfept.R
\name{read_generic_rfept}
\alias{read_generic_rfept}
\title{Read generic rfept files}
\usage{
read_generic_rfept(path = "", start = NULL)
}
\arguments{
\item{path}{The rfept file's path.}

\item{start}{A character with the date to start the comparison. The default is the minimum
value found in the data. The date has to be in this format: "yyyy.semester".
Ex.: "2019.1" or "2019.2".}
}
\value{
A data frame.
}
\description{
If your institute is not fully integrated with the package, you can transform your
academic database into a generic layout and use \code{read_rfept()} normally. See
Details to understand the requisites.
}
\details{
To transform your database in a generic rfept layout, follow these
requirements:
\itemize{
\item Rename your columns to: NO_ALUNO, NU_CPF, CO_MATRICULA, NO_STATUS_MATRICULA,
NO_CURSO, DT_INICIO_CURSO, NO_CAMPUS and NO_COTA;
\item All variables should be inherited to string class;
\item The CPF's should be in xxx.xxx.xxx-xx format;
\item The beginning date should be in yyyy.s format. Ex.: 2020.1. Use 1 for first
semester and 2 for second.
\item Convert the student's status to valid name in Sistec, use: ABANDONO, EM_CURSO,
CONCLUÍDA, DESLIGADO, INTEGRALIZADA, REPROVADA and TRANSF_EXT;
\item Save your data in a single file in csv format separated by comma and use
latin1 encoding. Semicolons separators and UTF-8 enconding are also available.
}
}
\examples{
 
# this dataset is not a real one. It is just for test purpose.
rfept <- read_generic_rfept(system.file("extdata/examples/generic_rfept",
                                        package = "sistec"))

rfept
}
