% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rfept.R
\name{read_rfept}
\alias{read_rfept}
\title{Identify and read academic registration}
\usage{
read_rfept(path = "")
}
\arguments{
\item{path}{The file's path to Qacademico or Sigaa folder.}
}
\value{
A data frame.
}
\description{
The \code{read_rfept()} is a wrapper around \code{read_qacademico()} and \code{read_sigaa()}. Now
you just need to specify the folder path and \code{read_rfept()} identifies if it is a
qacademico or sigaa file and then read it.
}
\details{
By now, this function only supports qacademico and sigaa.
}
\examples{
# these datasets are not a real ones. It is just for test purpose.

qacademico <- read_rfept(system.file("extdata/examples/qacademico",
                                     package = "sistec")) 
                                     
sigaa <- read_rfept(system.file("extdata/examples/sigaa",
                                 package = "sistec"))
                                 
class(qacademico)
class(sigaa)                                 
}
