% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sigaa.R
\name{read_sigaa}
\alias{read_sigaa}
\title{Read Sigaa files}
\usage{
read_sigaa(path = "")
}
\arguments{
\item{path}{The Sigaa file's path.}
}
\value{
A data frame.
}
\description{
This function support two kinds of schemas: from the api and the website. See Details
if you need help to download the Qacademico data.
}
\details{
AINDA FALTA FAZER PARA O SIGAA. To download the student's data, go to your proper account on Qacademico and
follow:
\itemize{
\item "Relatorio de Alunos" -> "Listagem de Alunos" (choose year and period)
\item Click on "visualizar"
\item Using F10 shortcut and save in .csv format
\item Rename the including year and period (example2020_1.csv)
}

Be sure that your data has the variables: "Matricula", "Nome", "Situacao Matricula",
"Curso", "Cpf", "Instituicao", "Per. Letivo Inicial".
}
