% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sistec.R
\name{read_sistec}
\alias{read_sistec}
\title{Read sistec files}
\usage{
read_sistec(path = "")
}
\arguments{
\item{path}{The sistec file's path.}
}
\value{
A data frame
}
\description{
This function support two kinds of schemas: from the api (setec) and from the website.
}
\examples{

# this dataset is not a real one. It is just for test purpose.
sistec <- read_sistec(system.file("extdata/examples/sistec",
                                  package = "sistec"))

sistec

}
