% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sistec_students.R
\name{read_sistec_students}
\alias{read_sistec_students}
\title{Read students

I'll write this in another time}
\usage{
read_sistec_students(path = "")
}
\arguments{
\item{path}{The student folder's path}
}
\description{
Read students

I'll write this in another time
}
