% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sistec_app.R
\name{sistec_app}
\alias{sistec_app}
\title{Sistec web application}
\usage{
sistec_app(
  output_path = NULL,
  output_folder_name = "Sistec_app",
  max_file_size = 100
)
}
\arguments{
\item{output_path}{The folder where you want to save the results.}

\item{output_folder_name}{The folder's name you want to save the results.}

\item{max_file_size}{The maximum file size in megabytes.}
}
\value{
A web application.
}
\description{
This is the web application using the sistec package. It was created to ease the
work using the package, but you can have the same results reading the
files (sistec::read_*()), comparing the results (sistec::compare_sistec()) and write
the outputs (sistec::write_output())
}
