% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sistec.R
\name{compare_sistec}
\alias{compare_sistec}
\title{Comparison between sistec and a student registration dataset}
\usage{
compare_sistec(x, sistec, write_output_path = NULL, institute = "Sistec")
}
\arguments{
\item{x}{A data frame read by \verb{sistec::read_*()}functions.}

\item{sistec}{A Sistec data frame read by \code{sistec::read_sistec()}function.}

\item{output_path}{Specify a path if you want to save the comparison. The results will be
saved in a folder named by \code{output_folder_name} parameter. The files created will be saved in
xlsx format.}

\item{output_folder_name}{The folder name where the comparison will be saved. If not specified,
the folder name will be "Sistec". This parameter is useless if \code{output_path} was not specified.}
}
\description{
Compare and save the student situation and the inconsistecies in the datasets.
}
