% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ciclo.R
\name{read_ciclo}
\alias{read_ciclo}
\title{Read ciclo

I'll write this in another time}
\usage{
read_ciclo(path = "")
}
\arguments{
\item{path}{The ciclo folder's path}
}
\description{
Read ciclo

I'll write this in another time
}
