% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\alias{read_sistec}
\alias{read_qacademico}
\title{Read sistec and qacademico files}
\usage{
read_sistec(path = "extdata")

read_qacademico(path = "extdata")
}
\arguments{
\item{path}{The sistec file's path.}
}
\description{
These functions support two kinds of schemas: from the api and from the website.
}
