% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_linked_courses.R
\name{read_linked_courses}
\alias{read_linked_courses}
\title{Read linked courses files}
\usage{
read_linked_courses(path = "")
}
\arguments{
\item{path}{The linked courses file's path.}
}
\value{
A data frame.
}
\description{
This function reads a .xlsx file with the linked courses between Sistec and the academic
registration.  The columns must be in this order: INICIO, CICLO, CURSO_SISTEC,
CURSO_RFEPT and CAMPUS. The date in INICIO column must be in yyyy.period.
Ex.: 2019.1 or 2019.2. .
}
\examples{
 
linked_courses <- read_linked_courses(system.file("extdata/examples/linked_courses",
                                                  package = "sistec"))

linked_courses
}
