% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qacademico.R
\name{read_qacademico}
\alias{read_qacademico}
\title{Read Qacademico files}
\usage{
read_qacademico(path = "")
}
\arguments{
\item{path}{The Qacademico file's path.}
}
\value{
A data frame
}
\description{
This function support two kinds of schemas: from the api and from the website.
}
\examples{

# this dataset is not a real one. It is just for test purpose.
qacademico <- read_qacademico(system.file("extdata/examples/qacademico",
                                          package = "sistec"))

qacademico
}
