% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sistec_students.R
\name{read_sistec_students}
\alias{read_sistec_students}
\title{Read students

This function provides support to read sistec students files that can
be extarct from sistec website.}
\usage{
read_sistec_students(path = "", start = NULL)
}
\arguments{
\item{path}{The folder's path to students files.
@param start A character with the date to start the comparison. The default is the minimum
value found in the data. The date has to be in this format: "yyyy.semester".
Ex.: "2019.1" or "2019.2".}
}
\description{
Read students

This function provides support to read sistec students files that can
be extarct from sistec website.
}
