% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sistec_app.R
\name{sistec_app}
\alias{sistec_app}
\title{Run Sistec_app}
\usage{
sistec_app(output_folder_name = "Sistec_app")
}
\arguments{
\item{output_folder_name}{The folder's name you want to save the results.}
}
\description{
This is the web application using the sistec package. It was created to ease the
work using the package, but you can have the same results reading the
files (sistec::read_*()), comparing the results (sistec::compare_sistec()) and write
the outputs (sistec::write_output())
}
