% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Save the comparison results}
\usage{
write_output(x, output_path = NULL, output_folder_name = "ARIA")
}
\arguments{
\item{x}{A list returned by \code{compare_sistec()}.}

\item{output_path}{The folder where you want to save the results.}

\item{output_folder_name}{The folder's name you want to save the results.}
}
\value{
None.
}
\description{
You can use this function to save the results saparated by campus. The resutls will be
saved in .xlsx format.
}
